Add files via upload
I am uploading the python code which I used to clean and reorganize the Food-World-Cup dataset that I used from the below mentioned link:

https://github.com/fivethirtyeight/data/blob/master/food-world-cup/food-world-cup-data.csv 

The above dataset was not fit for use in the Vega-Lite and it's scope was thus limited. 

From the new dataset that I created, I was able to generate very attractive and meaningful visualization which I uploaded on my github pages.

I am also uploading the old data set which I sourced from the below link and is not my property and I am only uploading for the purpose of comparison. The name of the old file that I am uploading is Food.csv